refactor: add custom list of dao token logos
